fix(launch): cast STILL_ACTIVE to u32 in process_alive on Windows
windows-sys 0.61 defines STILL_ACTIVE as an i32 (NTSTATUS), while
GetExitCodeProcess writes to a u32. The comparison in process_alive()
fails to compile on Windows (E0308: mismatched types).

+++ src/launch/mod.rs
@@ -290,7 +290,7 @@ pub(crate) fn process_alive(pid: u32) -> bool {
         let mut exit_code = 0u32;
         let ok = unsafe { GetExitCodeProcess(handle, &mut exit_code) } != 0;
         unsafe { CloseHandle(handle) };
-        ok && exit_code == STILL_ACTIVE
+        ok && exit_code == STILL_ACTIVE as u32
     }
     #[cfg(not(any(unix, windows)))]
     {
